Fix for speed controller

+++ ui/coord/CalxCoordArcCtrl.cpp
@@ -38,9 +38,9 @@ namespace CalXUI {
                                       wxDefaultSize, wxSP_ARROW_KEYS, INT_MIN, INT_MAX, 0, wxGetApp().getUnitPrecision());
         this->speed = new wxSpinCtrlDouble(this, wxID_ANY, wxEmptyString, wxDefaultPosition,
                                      wxDefaultSize, wxSP_ARROW_KEYS, 0,
-                                     (int) wxGetApp().getSystemManager()->getConfiguration()->
+                                     wxGetApp().getSystemManager()->getConfiguration()->
                                         getEntry("ui")->getReal("unit_speed", 4000.0),
-                                     (int) wxGetApp().getSystemManager()->getConfiguration()->
+                                     wxGetApp().getSystemManager()->getConfiguration()->
                                         getEntry("ui")->getReal("unit_speed", 4000.0), wxGetApp().getSpeedPrecision());
 		this->divisor = new wxChoice(this, wxID_ANY);
 		this->divisor->Append("1");
